made the Joi check in a separate file
